refactor: extract shared request handler from entry-prod
Moves API routing + SSR rendering logic into request-handler.ts
so it can be reused by different deployment adapters.
